+++ contrib/pyln-proto/pyln/proto/message/Makefile
@@ -1,0 +1,4 @@
+#! /usr/bin/make
+
+refresh:
+	for d in bolt*; do $(MAKE) -C $$d; done
