+++ contrib/pyln-proto/pyln/proto/message/bolt1/__init__.py
@@ -1,0 +1,16 @@
+from .csv import csv
+from .bolt import namespace
+import sys
+
+__version__ = '0.0.1'
+
+__all__ = [
+    'csv',
+    'namespace',
+]
+
+mod = sys.modules[__name__]
+for d in namespace.subtypes, namespace.tlvtypes, namespace.messagetypes:
+    for name in d:
+        setattr(mod, name, d[name])
+        __all__.append(name)
