+++ contrib/pyln-proto/pyln/proto/message/bolt1/bolt.py
@@ -1,0 +1,5 @@
+from pyln.proto.message import MessageNamespace
+from .csv import csv
+
+
+namespace = MessageNamespace(csv_lines=csv)
