--- contrib/pyln-proto/pyln/proto/message/bolts/__init__.py
@@ -1,10 +1,0 @@
-from .bolts import bolt_01_csv, bolt_02_csv, bolt_04_csv, bolt_07_csv
-
-__version__ = '0.0.1'
-
-__all__ = [
-    "bolt_01_csv",
-    "bolt_02_csv",
-    "bolt_04_csv",
-    "bolt_07_csv",
-]
