+++ contrib/pyln-proto/setup.py
@@ -17,7 +17,7 @@ setup(name='pyln-proto',
       author='Christian Decker',
       author_email='decker.christian@gmail.com',
       license='MIT',
-      packages=['pyln.proto', 'pyln.proto.message', 'pyln.proto.message.bolts'],
+      packages=['pyln.proto', 'pyln.proto.message', 'pyln.proto.message.bolts', 'pyln.proto.message.bolt1'],
       scripts=[],
       zip_safe=True,
       install_requires=requirements)
