+++ contrib/pyln-proto/tests/test_bolt2.py
@@ -1,0 +1,8 @@
+#! /usr/bin/python3
+from pyln.proto.message import MessageNamespace
+import pyln.proto.message.bolt2 as bolt2
+
+
+# FIXME: more tests
+def test_bolt_02_csv():
+    MessageNamespace(bolt2.csv)
